Refactor: Update SEO, UI, and dependencies
This commit updates meta tags, improves UI elements, and adjusts package dependencies for better performance and SEO.
