+++ index.html
@@ -3,8 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-<<<<<<< HEAD
-    <title>Zion - The Future of Tech & AI Marketplace</title>
+    <title>Zion Tech Group - Innovative IT Solutions & AI Services</title>
     <meta name="description" content="The world's first free marketplace dedicated to high-tech and artificial intelligence. Publish, connect, and thrive — all one place." />
     <meta name="author" content="Zion Tech Group" />
     <meta name="keywords" content="AI marketplace, tech services, artificial intelligence, IT services, AI developers, technology marketplace, AI solutions, machine learning, digital transformation" />
@@ -35,7 +34,7 @@
     <link rel="canonical" href="https://ziontechgroup.com/" />
     
     <!-- Favicon -->
-    <link rel="icon" href="https://drive.google.com/uc?export=view&id=0B0iuzhpa3pD7X0RzZ2lmclN3Ymc" />
+    <link rel="icon" href="/favicon.ico" />
     
     <!-- Apple Touch Icon -->
     <link rel="apple-touch-icon" href="https://drive.google.com/uc?export=view&id=0B0iuzhpa3pD7X0RzZ2lmclN3Ymc" />
@@ -56,22 +55,12 @@
         ]
       }
     </script>
-=======
-    <title>Zion Tech Group - Innovative IT Solutions & AI Services</title>
-    <link rel="icon" href="/favicon.ico" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Orbitron:wght@400;600&display=swap"
-      rel="stylesheet"
-    />
->>>>>>> 8b13c47f5ac42b659a51e0b68814bcab5e74cc20
   </head>
 
   <body>
     <div id="root"></div>
     <!-- IMPORTANT: DO NOT REMOVE THIS SCRIPT TAG OR THIS VERY COMMENT! -->
     <script src="https://cdn.gpteng.co/gptengineer.js" type="module"></script>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/main.js"></script>
   </body>
 </html>
